[Digitizing] Move tools logic in a module
